Move image files and add outline of tasks
